pyproject.toml
[project]
name = "nocodile-backend"
version = "0.1.0"
requires-python = "~=3.9"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "brokenaxes>=0.6.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "fastapi==0.115.11",
    "h11==0.14.0",
    "h5py>=3.13.0",
    "idna==3.10",
    "keras>=3.9.0; platform_system != 'Windows'",
    "keras~=2.10; platform_system == 'Windows'",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2; platform_system != 'Windows'",
    "numpy<2; platform_system == 'Windows'",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "tensorflow>=2.19.0; platform_system != 'Windows'",
    "tensorflow==2.10.0; platform_system == 'Windows'",
    "tensorflow-io-gcs-filesystem==0.27.0; platform_system == 'Windows'",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "orjson>=3.10.16",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
